requirements.txt
python-shogi
cshogi
torch
numpy
pandas
matplotlib